+++ src/main.js
@@ -3,7 +3,7 @@ import upperFirst from 'lodash/upperFirst'
 import camelCase from 'lodash/camelCase'
 import App from './App.vue'
 import router from './router'
-import store from './store'
+import store from './store/store'
 import BaseIcon from '@/components/BaseIcon'
 
 Vue.component('BaseIcon', BaseIcon)
