+++ src/store/store.js
@@ -1,0 +1,24 @@
+import Vue from 'vue'
+import Vuex from 'vuex'
+import * as user from '@/store/modules/user.js'
+import * as event from '@/store/modules/event.js'
+
+Vue.use(Vuex)
+
+export default new Vuex.Store({
+  modules: {
+    user,
+    event
+  },
+  state: {
+    categories: [
+      'sustainability',
+      'natural',
+      'animal welfare',
+      'housing',
+      'education',
+      'food',
+      'community'
+    ]
+  }
+})
