+++ src/views/EventCreate.vue
@@ -17,25 +17,37 @@
       <h3>Name & describe your event</h3>
       <div class="field">
         <label>Title</label>
-        <input v-model="event.title" type="text" placeholder="Add an event title"/>
+        <input
+          v-model="event.title"
+          type="text"
+          placeholder="Add an event title"
+        />
       </div>
 
       <div class="field">
         <label>Description</label>
-        <input v-model="event.description" type="text" placeholder="Add a description"/>
+        <input
+          v-model="event.description"
+          type="text"
+          placeholder="Add a description"
+        />
       </div>
 
       <h3>Where is your event?</h3>
-      <div class="field"> 
+      <div class="field">
         <label>Location</label>
-        <input v-model="event.location" type="text" placeholder="Add a location"/>
+        <input
+          v-model="event.location"
+          type="text"
+          placeholder="Add a location"
+        />
       </div>
 
       <h3>When is your event?</h3>
 
       <div class="field">
         <label>Date</label>
-        <datepicker v-model="event.date" placeholder="Select a date"/>
+        <datepicker v-model="event.date" placeholder="Select a date" />
       </div>
 
       <div class="field">
@@ -45,7 +57,7 @@
         </select>
       </div>
 
-      <input type="submit" class="button -fill-gradient" value="Submit"/>
+      <input type="submit" class="button -fill-gradient" value="Submit" />
     </form>
   </div>
 </template>
@@ -71,7 +83,7 @@ export default {
   methods: {
     createEvent() {
       this.$store
-        .dispatch('createEvent', this.event)
+        .dispatch('event/createEvent', this.event)
         .then(() => {
           this.$router.push({
             name: 'event-show',
@@ -84,7 +96,7 @@ export default {
         })
     },
     createFreshEventObject() {
-      const user = this.$store.state.user
+      const user = this.$store.state.user.user
       const id = Math.floor(Math.random() * 10000000)
       return {
         id: id,
